Hasta donde pude. Ya veré luego los detalles

+++ src/main/java/com/example/tarea4_grupo2/dto/TiempoMedio_ClienteDTO.java
@@ -2,5 +2,5 @@ package com.example.tarea4_grupo2.dto;
 
 public interface TiempoMedio_ClienteDTO {
     String getNombrerestaurante();
-    int getTiempopromedio();
+    Double getTiempopromedio();
 }
